+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Changelog
 * 2.1.1:
+    * New: Tools to remove zombie variations and meta.
     * Fix: Remove debug method from, create variations from all attributes.
     * Fix: Hide out of stock products option.
     * Fix: Adding related products to cart.